+++ src/Client/ClientCore/Systems/ClientNetwork/ClientNetworkSystem.cs
@@ -42,12 +42,9 @@ public sealed class ClientNetworkSystem : ISystem, IDisposable
 
         eventDescriptions.RegisterNullEvent(EventId.Client_Network_Connected);
         eventDescriptions.RegisterNullEvent(EventId.Client_Network_ConnectionLost);
-        eventDescriptions.RegisterNullEvent(EventId.Client_Network_RegisterSuccess);
         eventDescriptions.RegisterEvent<BeginConnectionEventDetails>(EventId.Client_Network_BeginConnection);
         eventDescriptions.RegisterEvent<ErrorEventDetails>(EventId.Client_Network_ConnectionAttemptFailed);
         eventDescriptions.RegisterEvent<ErrorEventDetails>(EventId.Client_Network_LoginFailed);
-        eventDescriptions.RegisterEvent<RegisterAccountEventDetails>(EventId.Client_Network_RegisterAccount);
-        eventDescriptions.RegisterEvent<ErrorEventDetails>(EventId.Client_Network_RegisterFailed);
 
         eventLoop.RegisterSystem(this);
     }
@@ -63,7 +60,6 @@ public sealed class ClientNetworkSystem : ISystem, IDisposable
     {
         EventId.Client_Network_ConnectionLost,
         EventId.Client_Network_BeginConnection,
-        EventId.Client_Network_RegisterAccount,
         EventId.Core_WorldManagement_Subscribe,
         EventId.Core_WorldManagement_Unsubscribe
     };